A selected piece can now also be rotated without the +/- buttons. Just drag with mouse around it!
Details:
Enter Level Editor with ’e’ key.
Create a piece (e.g. an Asteroid) by placing it somewhere.
Select it by clicking on it (a selected piece has a box drawn around
it).
The piece can now be moved and rotated with mouse-click-and-drag. If
clicked within the box the piece gets moved around, if outside the box
the piece is rotated.

Click the selected piece again to deselect it, or just click anywhere
outside the box.